feat: new `IBuildProvider` and `ITransformProvider` to unify esbuild provider options (#286)
BREAKING CHANGE: `buildFn`, `transformFn` as wel as `esbuildBinaryPath` and `esbuildModulePath` have been replaced with a unified interface. Configure the default `EsbuildProvider` for more options or provide a custom implementation as an escape hatch.

BREAKING CHANGE: `props.copyDir` now removes the destination directory and any containing files if it exists. This will only affect edge cases when an asset has already been bundled and re-bundling is forced. Previously the destination directory would not have been cleared out, leading to outdated files being included in the bundle or an error message in same cases. The previous behavior was not correct and cannot be restored.